NERDd: Watchdog added to UM, new way of program termination
UpdateManage now has a watchdog which checks if all worker threads are
running and restarts them if some die.
There's a limit on number of restarts, after the limit is reached, program
is terminated.
This required to change a way the program can be terminated - now any
module and any thread can terminate the program gracefully by releasing
the g.daemon_stop_lock lock.