First try of BGP Ranking
I have encountered some problems with importing modules and I had
to change them a little bit. I guess this is happening because I
do not have se the path as it set on the machine with NERD. However,
this should work aswell.
The second problem is while importing bgpranking_web module, with
the same issue (cannot import api even though it is in the same
folder as the __init__.py file).
Or is it some beginner mistake?