Rep. score of BGP prefixes, API endpoint to get the worst ones
Rep. score of BGP prefixes is computed as average of rep. score of IPs addresses within in. It's recomputed once an hour by an external script (it's not the best solution, but it was needed to implement it quickly).
New API endpoint was created to get a list of the worst prefixes.