TaskQueueReader: added locking into ack_msg to avoid a random freeze bug
Pika is not thread-safe and Ack'ing can be called from worker-threads, so locking is necessary.

Also, prefetch count increased to improve performance (and to avoid a situation when normal messages are fetched although priority ones are in the RMQ, since local priority queue was just emptied and no new ones were received yet).