web: Add Hilbert map (#23)
Add page with visualization of IPv4 address space on NERD web. Visualization uses Hilbert curve to display IPv4 address space. Pixel on visualization represents a certain network and its value is sum of reputation scores of IP addresses in given network.

It uses iframe for displaying visualization. It is probably a temporary solution but it is the quickest one due to the remaining time.