blacklists2redis: Fixed a bug in storing prefix blacklists
The current way of storing prefix blacklists in Redis didn't work with /32 prefixes - in such case, the order of beginning and end entries was undefined (same "score"), which could lead to a case where the end of a prefix preceded its beginning, which resulted in IPs not appearing on blacklist marked as listed and vice versa.
Fixed by adding 0.5 to score of the prefix end entry, so it's always sorted after the beginning.