web: timestamped static files
In order to solve problems with browsers using old cached CSS and JS files when a new version is deployed, links to all CSS and JS files now contain timestamp of the file modification, thus forcing the browser to download it again. Apache configuration contains a RewriteRule which removes the timestamp from URL, so it points to the file itself.

Inspired by https://stackoverflow.com/a/118886/12777430