update_manager: Limit waiting on worker threads on exit
When program stops, the main thread waits for all the worker threads to finish their work. However, when the stop is initiated because some worker fails, it can wait infinitely. Therefore, a this change adds a limit to wait at most 2 minutes and then stop the program even if some threads are still "alive" (but probably stuck).