Shodan: Don't update short-lived addresses
Due to rate-limit of InternetDB API, we need to lower the number of requests, so we now skip periodic weekly updates of IPs that are not active for a long time (don't have "long-active" ttl tag).
Most of the IPs in the database have just a few reports for a short period of time, these are not very important and get removed after two week. At the same time, there's a lot of them, so we can save a lot by not updating them.